continue separating duties
