+++ modules/gob-web/modules/area-editor/index.js
@@ -1,6 +1,3 @@
 // @flow
 
-import Loadable from 'react-loadable'
-import {LoadingComponent} from '../../components/loading-comp'
-
-export {Editor} from './editor'
+export {Controller as AreaEditor} from './controller'
